django_echoshield/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0
httpx>=0.25.0
celery>=5.3.0
redis>=5.0.0
folium>=0.15.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
